Add solution for problem `D - Epic Transformation` Link: https://codeforces.com/contest/1506/problem/D Tags: constructive algorithms, data structures, greedy, *1400 Ref: https://codeforces.com/contest/1506/submission/137581650

+++ codeforces/1506/D.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 #define dbg(...)
 #endif
-using ll = int;
+using ll = int64_t;
 auto sz = [](const auto &container) -> ll { return container.size(); };
 #define all(x) (x).begin(), (x).end()
 template <class T, class U = T>
